✨ Add existing-result-contributors query and handler
Introduce a new query to fetch existing result contributors for a TOC indicator: GetExistingResultContributorsToIndicatorsQuery and its handler. Added supporting services and utilities: loader (loads and validates toc/result records), mapper (maps and sums contributing_indicator values), types, and a ContributorsRoleResolverService (resolves user general and per-result roles). Registered the new providers in results-framework-reporting.module and updated the service spec to include them. Refactored ResultsFrameworkReportingService to delegate the previous inline logic to the new handler (removed related imports and duplicated code).
