+++ onecgiar-pr-server/src/api/results-framework-reporting/application/commands/create-result-from-framework/create-result-from-framework.command.ts
@@ -1,5 +1,5 @@
-import { CreateResultsFrameworkResultDto } from '../../dto/create-results-framework.dto';
-import { TokenDto } from '../../../../shared/globalInterfaces/token.dto';
+import { CreateResultsFrameworkResultDto } from '../../../dto/create-results-framework.dto';
+import { TokenDto } from '../../../../../shared/globalInterfaces/token.dto';
 
 export class CreateResultFromFrameworkCommand {
   constructor(
