+++ onecgiar-pr-server/src/api/results-framework-reporting/application/commands/create-result-from-framework/framework-result-toc-indicators.service.ts
@@ -1,8 +1,8 @@
 import { HttpStatus, Injectable } from '@nestjs/common';
-import { ResultsFrameworkTocIndicatorDto } from '../../dto/create-results-framework.dto';
-import { AoWBilateralRepository } from '../../../results/results-toc-results/repositories/aow-bilateral.repository';
-import { ResultsTocResultIndicatorsRepository } from '../../../results/results-toc-results/repositories/results-toc-results-indicators.repository';
-import { ResultsTocTargetIndicatorRepository } from '../../../results/results-toc-results/repositories/result-toc-result-target-indicator.repository';
+import { ResultsFrameworkTocIndicatorDto } from '../../../dto/create-results-framework.dto';
+import { AoWBilateralRepository } from '../../../../results/results-toc-results/repositories/aow-bilateral.repository';
+import { ResultsTocResultIndicatorsRepository } from '../../../../results/results-toc-results/repositories/results-toc-results-indicators.repository';
+import { ResultsTocTargetIndicatorRepository } from '../../../../results/results-toc-results/repositories/result-toc-result-target-indicator.repository';
 import { throwReportingFrameworkError } from '../../utils/reporting-framework-error.util';
 
 type IndicatorNumberTargetValue = string | number | null;
