+++ onecgiar-pr-server/src/api/results-framework-reporting/application/queries/get-existing-result-contributors/get-existing-result-contributors.query.ts
@@ -1,0 +1,9 @@
+import { TokenDto } from '../../../../../shared/globalInterfaces/token.dto';
+
+export class GetExistingResultContributorsToIndicatorsQuery {
+  constructor(
+    public readonly user: TokenDto,
+    public readonly resultTocResultId: string | number,
+    public readonly tocResultIndicatorId: string,
+  ) {}
+}
