+++ onecgiar-pr-server/src/api/results-framework-reporting/results-framework-reporting.module.ts
@@ -23,11 +23,14 @@ import { ResultsByInstitutionsModule } from '../results/results_by_institutions/
 import { GeographicLocationModule } from './geographic-location/geographic-location.module';
 import { GeoScopeRoleModule } from './geo_scope_role/geo_scope_role.module';
 import { ReportingTocContextService } from './reporting-toc-context/reporting-toc-context.service';
-import { CreateResultFromFrameworkHandler } from './commands/create-result-from-framework/create-result-from-framework.handler';
-import { CreateFrameworkResultEntityService } from './commands/create-result-from-framework/create-framework-result-entity.service';
-import { LinkFrameworkResultTocService } from './commands/create-result-from-framework/link-framework-result-toc.service';
-import { FrameworkResultTocIndicatorsService } from './commands/create-result-from-framework/framework-result-toc-indicators.service';
-import { ApplyFrameworkResultAssociationsService } from './commands/create-result-from-framework/apply-framework-result-associations.service';
+import { CreateResultFromFrameworkHandler } from './application/commands/create-result-from-framework/create-result-from-framework.handler';
+import { CreateFrameworkResultEntityService } from './application/commands/create-result-from-framework/create-framework-result-entity.service';
+import { LinkFrameworkResultTocService } from './application/commands/create-result-from-framework/link-framework-result-toc.service';
+import { FrameworkResultTocIndicatorsService } from './application/commands/create-result-from-framework/framework-result-toc-indicators.service';
+import { ApplyFrameworkResultAssociationsService } from './application/commands/create-result-from-framework/apply-framework-result-associations.service';
+import { GetExistingResultContributorsToIndicatorsHandler } from './application/queries/get-existing-result-contributors/get-existing-result-contributors.handler';
+import { ExistingResultContributorsLoaderService } from './application/queries/get-existing-result-contributors/existing-result-contributors-loader.service';
+import { ContributorsRoleResolverService } from './application/queries/get-existing-result-contributors/contributors-role-resolver.service';
 
 @Module({
   imports: [
@@ -62,6 +65,9 @@ import { ApplyFrameworkResultAssociationsService } from './commands/create-resul
     LinkFrameworkResultTocService,
     FrameworkResultTocIndicatorsService,
     ApplyFrameworkResultAssociationsService,
+    GetExistingResultContributorsToIndicatorsHandler,
+    ExistingResultContributorsLoaderService,
+    ContributorsRoleResolverService,
   ],
   exports: [ReportingTocContextService],
 })
