+++ onecgiar-pr-server/src/api/results-framework-reporting/results-framework-reporting.service.spec.ts
@@ -19,11 +19,14 @@ import { ContributionToIndicatorResultsRepository } from '../contribution-to-ind
 import { ResultsTocTargetIndicatorRepository } from '../results/results-toc-results/repositories/result-toc-result-target-indicator.repository';
 import { ResultsByInstitutionsService } from '../results/results_by_institutions/results_by_institutions.service';
 import { ReportingTocContextService } from './reporting-toc-context/reporting-toc-context.service';
-import { CreateResultFromFrameworkHandler } from './commands/create-result-from-framework/create-result-from-framework.handler';
-import { CreateFrameworkResultEntityService } from './commands/create-result-from-framework/create-framework-result-entity.service';
-import { LinkFrameworkResultTocService } from './commands/create-result-from-framework/link-framework-result-toc.service';
-import { FrameworkResultTocIndicatorsService } from './commands/create-result-from-framework/framework-result-toc-indicators.service';
-import { ApplyFrameworkResultAssociationsService } from './commands/create-result-from-framework/apply-framework-result-associations.service';
+import { CreateResultFromFrameworkHandler } from './application/commands/create-result-from-framework/create-result-from-framework.handler';
+import { CreateFrameworkResultEntityService } from './application/commands/create-result-from-framework/create-framework-result-entity.service';
+import { LinkFrameworkResultTocService } from './application/commands/create-result-from-framework/link-framework-result-toc.service';
+import { FrameworkResultTocIndicatorsService } from './application/commands/create-result-from-framework/framework-result-toc-indicators.service';
+import { ApplyFrameworkResultAssociationsService } from './application/commands/create-result-from-framework/apply-framework-result-associations.service';
+import { GetExistingResultContributorsToIndicatorsHandler } from './application/queries/get-existing-result-contributors/get-existing-result-contributors.handler';
+import { ExistingResultContributorsLoaderService } from './application/queries/get-existing-result-contributors/existing-result-contributors-loader.service';
+import { ContributorsRoleResolverService } from './application/queries/get-existing-result-contributors/contributors-role-resolver.service';
 
 const mockClarisaInitiativesRepository = {
   findOne: jest.fn(),
@@ -205,6 +208,9 @@ describe('ResultsFrameworkReportingService', () => {
         LinkFrameworkResultTocService,
         FrameworkResultTocIndicatorsService,
         ApplyFrameworkResultAssociationsService,
+        GetExistingResultContributorsToIndicatorsHandler,
+        ExistingResultContributorsLoaderService,
+        ContributorsRoleResolverService,
       ],
     }).compile();
